--- a/lib/cast/number.js
+++ b/lib/cast/number.js
@@ -3,12 +3,11 @@
 const assert = require('assert');
 
 /*!
- * Given a value, cast it to a number, or throw a `CastError` if the value
+ * Given a value, cast it to a number, or throw an `Error` if the value
  * cannot be casted. `null` and `undefined` are considered valid.
  *
  * @param {Any} value
- * @param {String} [path] optional the path to set on the CastError
- * @return {Boolean|null|undefined}
+ * @return {Number}
  * @throws {Error} if `value` is not one of the allowed values
  * @api private
  */