--- a/lib/helpers/query/cast$expr.js
+++ b/lib/helpers/query/cast$expr.js
@@ -68,7 +68,7 @@ const dateOperators = new Set([
 ]);
 
 module.exports = function cast$expr(val, schema, strictQuery) {
-  if (typeof val !== 'object' || val == null) {
+  if (typeof val !== 'object' || val === null) {
     throw new Error('`$expr` must be an object');
   }
 
@@ -123,10 +123,8 @@ function _castExpression(val, schema, strictQuery) {
 
 function _omitUndefined(val) {
   const keys = Object.keys(val);
-  for (const key of keys) {
-    if (val[key] === void 0) {
-      delete val[key];
-    }
+  for (let i = 0, len = keys.length; i < len; ++i) {
+    (val[keys[i]] === void 0) && delete val[keys[i]];
   }
 }
 
@@ -144,15 +142,14 @@ function castNumberOperator(val) {
 }
 
 function castIn(val, schema, strictQuery) {
-  let search = val[0];
-  let path = val[1];
+  const path = val[1];
   if (!isPath(path)) {
     return val;
   }
+  const search = val[0];
 
-  path = path.slice(1);
-  const schematype = schema.path(path);
-  if (schematype == null) {
+  const schematype = schema.path(path.slice(1));
+  if (schematype === null) {
     if (strictQuery === false) {
       return val;
     } else if (strictQuery === 'throw') {
@@ -166,12 +163,10 @@ function castIn(val, schema, strictQuery) {
     throw new Error('Path must be an array for $in');
   }
 
-  if (schematype.$isMongooseDocumentArray) {
-    search = schematype.$embeddedSchemaType.cast(search);
-  } else {
-    search = schematype.caster.cast(search);
-  }
-  return [search, val[1]];
+  return [
+    schematype.$isMongooseDocumentArray ? schematype.$embeddedSchemaType.cast(search) : schematype.caster.cast(search),
+    path
+  ];
 }
 
 // { $op: [<number>, <number>] }
@@ -268,14 +263,14 @@ function castComparison(val, schema, strictQuery) {
 }
 
 function isPath(val) {
-  return typeof val === 'string' && val.startsWith('$');
+  return typeof val === 'string' && val[0] === '$';
 }
 
 function isLiteral(val) {
-  if (typeof val === 'string' && val.startsWith('$')) {
+  if (typeof val === 'string' && val[0] === '$') {
     return false;
   }
-  if (typeof val === 'object' && val != null && Object.keys(val).find(key => key.startsWith('$'))) {
+  if (typeof val === 'object' && val !== null && Object.keys(val).find(key => key[0] === '$')) {
     // The `$literal` expression can make an object a literal
     // https://docs.mongodb.com/manual/reference/operator/aggregation/literal/#mongodb-expression-exp.-literal
     return val.$literal != null;